more refactoring

+++ mcs/mcs/include/position.h
@@ -3,6 +3,6 @@
 struct position_t
 {
 	i32 x;
-	i16 y;
+	i32 y;
 	i32 z;
 };
