+++ mcs/mcs/src/util/def.c
@@ -1,7 +1,7 @@
 #include "util/def.h"
 #include <stdlib.h>
 #include <assert.h>
-void *mcsalloc(u64 s)
+void *mcsalloc(i32 s)
 {
 	void* m = malloc(s);
 	if(m == 0)
@@ -12,7 +12,7 @@ void *mcsalloc(u64 s)
 	return m;
 }
 
-void *mcsrealloc(void *block, u64 s)
+void *mcsrealloc(void *block, i32 s)
 {
 	void* m = realloc(block, s);
 	if(m == 0)
